+++ packages/image-cloze-association/src/image-drop-target.jsx
@@ -3,48 +3,83 @@ import PropTypes from 'prop-types';
 import { DropTarget } from '@pie-lib/pie-toolbox/drag';
 import { withStyles } from '@material-ui/core/styles';
 import { color } from '@pie-lib/pie-toolbox/render-ui';
+import cx from 'classnames';
 
 import PossibleResponse from './possible-response';
 import c from './constants';
 
-const ImageDropTarget = ({
-  answers,
-  canDrag,
-  classes,
-  containerStyle,
-  draggingElement,
-  duplicateResponses,
-  onDragAnswerBegin,
-  onDragAnswerEnd,
-  showDashedBorder,
-  // dnd-related props
-  connectDropTarget,
-}) =>
-  connectDropTarget(
-    <div
-      className={`
-        ${classes.responseContainer}
-        ${showDashedBorder && !draggingElement.id ? classes.responseContainerDashed : ''}
-        ${draggingElement.id ? classes.responseContainerActive : ''}
-      `}
-      style={containerStyle}
-    >
-      {answers.length || (duplicateResponses && answers.length) ? (
-        <div className={classes.answers}>
-          {(answers || []).map((answer) => (
-            <PossibleResponse
-              canDrag={canDrag}
-              containerStyle={answer.isCorrect === undefined ? { borderWidth: 0 } : {}}
-              key={answer.id}
-              data={answer}
-              onDragBegin={() => onDragAnswerBegin(answer)}
-              onDragEnd={onDragAnswerEnd}
-            />
-          ))}
-        </div>
-      ) : null}
-    </div>,
-  );
+class ImageDropTarget extends React.Component {
+  state = {
+    shouldHaveSmallPadding: false
+  };
+
+  render() {
+    const {
+      answers,
+      canDrag,
+      classes,
+      containerStyle,
+      draggingElement,
+      duplicateResponses,
+      onDragAnswerBegin,
+      onDragAnswerEnd,
+      showDashedBorder,
+      responseAreaFill,
+      // dnd-related props
+      connectDropTarget,
+      answerChoiceTransparency
+    } = this.props;
+    const { shouldHaveSmallPadding } = this.state;
+
+    const containerClasses = cx(classes.responseContainer, {
+      [classes.responseContainerDashed]: showDashedBorder && !draggingElement.id,
+      [classes.responseContainerActive]: draggingElement.id,
+    });
+
+    const updatedContainerStyle = {
+      ...containerStyle,
+      ...(responseAreaFill ? { backgroundColor: responseAreaFill } : {})
+    };
+
+    return connectDropTarget(
+        <div
+            className={containerClasses}
+            style={updatedContainerStyle}
+            ref={ref => {
+              this.dropContainerHeight = ref?.getBoundingClientRect().height;
+
+              // If answers area is higher than the initial established response zone,
+              // then make sure to use a smaller padding.
+              // Only reset the state if needed to prevent maximum callstack exceeded
+              if (this.dropContainerHeight < this.dropContainerResponsesHeight && !shouldHaveSmallPadding) {
+                this.setState({ shouldHaveSmallPadding: true });
+              }
+            }}
+        >
+          {answers.length || (duplicateResponses && answers.length) ? (
+              <div
+                  className={classes.answers}
+                   ref={ref => {
+                     this.dropContainerResponsesHeight = ref?.getBoundingClientRect().height;
+                   }}
+              >
+                {(answers || []).map((answer) => (
+                    <PossibleResponse
+                        canDrag={canDrag}
+                        key={answer.id}
+                        data={answer}
+                        onDragBegin={() => onDragAnswerBegin(answer)}
+                        onDragEnd={onDragAnswerEnd}
+                        answerChoiceTransparency={answerChoiceTransparency}
+                        containerStyle={shouldHaveSmallPadding ? { padding: '2px'} : { padding: '6px 10px' }}
+                    />
+                ))}
+              </div>
+          ) : null}
+        </div>,
+    );
+  }
+}
 
 ImageDropTarget.propTypes = {
   answer: PropTypes.object,
@@ -56,6 +91,8 @@ ImageDropTarget.propTypes = {
   onDragAnswerEnd: PropTypes.func.isRequired,
   onDrop: PropTypes.func.isRequired,
   showDashedBorder: PropTypes.bool,
+  responseAreaFill: PropTypes.string,
+  answerChoiceTransparency: PropTypes.bool
 };
 
 ImageDropTarget.defaultProps = {
@@ -74,7 +111,9 @@ const styles = () => ({
     width: 'fit-content',
   },
   responseContainer: {
+    padding: '10px',
     position: 'absolute',
+    boxSizing: 'border-box'
   },
   responseContainerActive: {
     border: `2px solid ${color.text()}`,