+++ packages/image-cloze-association/src/root.jsx
@@ -5,7 +5,7 @@ import { CSSTransition, TransitionGroup } from 'react-transition-group';
 import { ShowRationale } from '@pie-lib/pie-toolbox/icons';
 import { color, Collapsible, PreviewPrompt, hasText } from '@pie-lib/pie-toolbox/render-ui';
 import { withStyles } from '@material-ui/core/styles';
-import {CorrectAnswerToggle} from '@pie-lib/pie-toolbox/correct-answer-toggle';
+import { CorrectAnswerToggle } from '@pie-lib/pie-toolbox/correct-answer-toggle';
 import Translator from '@pie-lib/pie-toolbox/translator';
 
 const { translator } = Translator;
@@ -30,7 +30,7 @@ const styles = (theme) => ({
   },
   rationale: {
     marginTop: theme.spacing.unit * 2,
-  }
+  },
 });
 
 class ImageClozeAssociationComponent extends React.Component {
@@ -89,11 +89,16 @@ class ImageClozeAssociationComponent extends React.Component {
   };
 
   filterPossibleAnswers = (possibleResponses, answer) => {
-    const index = possibleResponses.findIndex(response => response.value === answer.value);
-    return [
-      ...possibleResponses.slice(0, index),       // Elements before the found item
-      ...possibleResponses.slice(index + 1)       // Elements after the found item
-    ];
+    const index = possibleResponses.findIndex((response) => response.value === answer.value);
+
+    if (index >= 0) {
+      return [
+        ...possibleResponses.slice(0, index), // Elements before the found item
+        ...possibleResponses.slice(index + 1), // Elements after the found item
+      ];
+    }
+
+    return possibleResponses;
   };
 
   handleOnAnswerSelect = (answer, responseContainerIndex) => {
@@ -101,7 +106,8 @@ class ImageClozeAssociationComponent extends React.Component {
       model: { duplicateResponses },
       updateAnswer,
     } = this.props;
-    const { answers, possibleResponses, maxResponsePerZone } = this.state;
+    const { answers, maxResponsePerZone } = this.state;
+    let { possibleResponses } = this.state;
     let answersToStore;
 
     if (maxResponsePerZone === answers.filter((a) => a.containerIndex === responseContainerIndex).length) {
@@ -164,9 +170,7 @@ class ImageClozeAssociationComponent extends React.Component {
       answers: answersToStore,
       possibleResponses:
         // for single response per container remove answer from possible responses
-        duplicateResponses
-          ? possibleResponses
-          : this.filterPossibleAnswers(possibleResponses, answer)
+        duplicateResponses ? possibleResponses : this.filterPossibleAnswers(possibleResponses, answer),
     });
     updateAnswer(answersToStore);
   };
@@ -207,6 +211,7 @@ class ImageClozeAssociationComponent extends React.Component {
         disabled,
         duplicateResponses,
         image,
+        responseAreaFill,
         stimulus,
         responseCorrect,
         validation,
@@ -217,6 +222,7 @@ class ImageClozeAssociationComponent extends React.Component {
         rationale,
         language,
         uiStyle = {},
+        answerChoiceTransparency,
       },
     } = this.props;
     const {
@@ -245,8 +251,10 @@ class ImageClozeAssociationComponent extends React.Component {
       });
     }
 
-    const warningMessage = translator.t('imageClozeAssociation.reachedLimit_other',
-      { lng: language, count: maxResponsePerZone });
+    const warningMessage = translator.t('imageClozeAssociation.reachedLimit_other', {
+      lng: language,
+      count: maxResponsePerZone,
+    });
 
     let answersToShow =
       responseCorrect !== undefined ? getAnswersCorrectness(answers, validation, duplicateResponses) : answers;
@@ -273,7 +281,12 @@ class ImageClozeAssociationComponent extends React.Component {
 
         <PreviewPrompt defaultClassName={classes.stimulus} prompt={stimulus} />
 
-        <CorrectAnswerToggle show={showToggle} toggled={showCorrect} onToggle={this.toggleCorrect} language={language}/>
+        <CorrectAnswerToggle
+          show={showToggle}
+          toggled={showCorrect}
+          onToggle={this.toggleCorrect}
+          language={language}
+        />
 
         {showCorrect && showToggle ? (
           <InteractiveSection responseCorrect={true} uiStyle={uiStyle}>
@@ -288,6 +301,7 @@ class ImageClozeAssociationComponent extends React.Component {
               onDragAnswerEnd={this.handleOnDragEnd}
               responseContainers={responseContainers}
               showDashedBorder={showDashedBorder}
+              responseAreaFill={responseAreaFill}
             />
           </InteractiveSection>
         ) : (
@@ -303,6 +317,8 @@ class ImageClozeAssociationComponent extends React.Component {
               onDragAnswerEnd={this.handleOnDragEnd}
               responseContainers={responseContainers}
               showDashedBorder={showDashedBorder}
+              responseAreaFill={responseAreaFill}
+              answerChoiceTransparency={answerChoiceTransparency}
             />
 
             {maxResponsePerZoneWarning && <WarningInfo message={warningMessage} />}
@@ -313,6 +329,8 @@ class ImageClozeAssociationComponent extends React.Component {
               onAnswerRemove={this.handleOnAnswerRemove}
               onDragBegin={this.beginDrag}
               onDragEnd={this.handleOnDragEnd}
+              answerChoiceTransparency={answerChoiceTransparency}
+              customStyle={{ minWidth: image?.width || 'fit-content' }}
             />
           </InteractiveSection>
         )}