+++ packages/hotspot/CHANGELOG.md
@@ -3,6 +3,65 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [6.25.2](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot@6.25.1...@pie-element/hotspot@6.25.2) (2024-08-19)
+
+
+### Bug Fixes
+
+* update pie-lib PD-3855, PD-3841 ([29ed50e](https://github.com/pie-framework/pie-elements/commit/29ed50ec75c3deedfe765f776dc24a6ecd6af284))
+
+
+
+
+
+## [6.25.1](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot@6.25.0...@pie-element/hotspot@6.25.1) (2024-08-06)
+
+
+### Bug Fixes
+
+* update pie-lib ([6ed3619](https://github.com/pie-framework/pie-elements/commit/6ed3619e0c670165ab45518e5bdbb40586c5adf8))
+
+
+
+
+
+# [6.25.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot@6.24.0...@pie-element/hotspot@6.25.0) (2024-08-06)
+
+
+### Features
+
+* **hotspot:** Add selectedHotspotColor and hoverOutlineColor support PD-3852 ([e6f8ed8](https://github.com/pie-framework/pie-elements/commit/e6f8ed887a0ceaaaaf11d78d5a736c723f19600c))
+* **hotspot:** Add selectedHotspotColor and hoverOutlineColor support PD-3852 ([422a426](https://github.com/pie-framework/pie-elements/commit/422a426e0189e7b4110216a8ef3599bbf74db17f))
+* update pie-lib PD-3996, PD-3792, Pd-3791, PD-3841, PD-3956, PD-3569, PD-3855 ([6343d1b](https://github.com/pie-framework/pie-elements/commit/6343d1b00c5a4a3d88de70ed13a8aa5c1a43002d))
+* **hotspot:** Add selectedHotspotColor and hoverOutlineColor support PD-3852 ([fd148d4](https://github.com/pie-framework/pie-elements/commit/fd148d423dfca651b8c420d2e2913c7df4dda0b7))
+
+
+
+
+
+# [6.24.0](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot@6.23.6...@pie-element/hotspot@6.24.0) (2024-07-26)
+
+
+### Features
+
+* update pie-lib/pie-toolbox PD-2451, PD-3950, PD-3845 ([278d705](https://github.com/pie-framework/pie-elements/commit/278d7052dc7c17796ba196a2e397739e2e882b7f))
+* update pie-lib/pie-toolbox PD-3944, PD-3946, Pd-2450 ([ac34d93](https://github.com/pie-framework/pie-elements/commit/ac34d93fcc97e79cacddcd38c7509f3c96b69976))
+
+
+
+
+
+## [6.23.6](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot@6.23.5...@pie-element/hotspot@6.23.6) (2024-07-19)
+
+
+### Bug Fixes
+
+* udate pie-lib/pie-toolbox to revert PD-3630 ([7858313](https://github.com/pie-framework/pie-elements/commit/785831349bb7830a3df8f35971958ebf616527f3))
+
+
+
+
+
 ## [6.23.5](https://github.com/pie-framework/pie-elements/compare/@pie-element/hotspot@6.23.4...@pie-element/hotspot@6.23.5) (2024-07-17)